+++ src/Products/PloneMeeting/content/advice.py
@@ -4,6 +4,7 @@ from AccessControl import ClassSecurityInfo
 from AccessControl import Unauthorized
 from App.class_init import InitializeClass
 from collective.contact.plonegroup.utils import get_organization
+from dexterity.localrolesfield.field import LocalRoleField
 from imio.history.utils import getLastWFAction
 from imio.prettylink.interfaces import IPrettyLink
 from persistent.list import PersistentList
@@ -37,7 +38,8 @@ class IMeetingAdvice(IMeetingContent):
     """
         MeetingAdvice schema
     """
-    advice_group = schema.Choice(
+
+    advice_group = LocalRoleField(
         title=_(u'Group'),
         description=_(u"Choose a group."),
         vocabulary=u'Products.PloneMeeting.content.advice.advice_group_vocabulary',
@@ -313,10 +315,12 @@ class AdviceGroupVocabulary(object):
         # while adding an advice, the context is his parent, aka a MeetingItem
         alterable_advice_org_uids = []
         if context.meta_type == 'MeetingItem':
-            alterable_advice_org_uids = [org_uid for org_uid, org_title in context.getAdvicesGroupsInfosForUser()[0]]
+            alterable_advice_org_uids = [org_uid for org_uid, org_title in
+                                         context.getAdvicesGroupsInfosForUser(compute_to_edit=False)[0]]
         # take into account groups for which user can edit an advice
         elif context.portal_type in advicePortalTypeIds:
-            alterable_advice_org_uids = [org_uid for org_uid, org_title in context.getAdvicesGroupsInfosForUser()[1]]
+            alterable_advice_org_uids = [org_uid for org_uid, org_title in
+                                         context.getAdvicesGroupsInfosForUser(compute_to_add=False)[1]]
             # make sure advice_group selected on advice is in the vocabulary
             if context.advice_group not in alterable_advice_org_uids:
                 alterable_advice_org_uids.append(context.advice_group)